Make plugin python3 compatible

+++ src/omero_cli_metadata.py
@@ -151,7 +151,7 @@ class MetadataControl(BaseControl):
         pixelsize = parser.add(sub, self.pixelsize)
 
         populate.add_argument("--batch",
-                              type=long,
+                              type=int,
                               default=1000,
                               help="Number of objects to process at once")
         self._add_wait(populate)
@@ -195,12 +195,12 @@ class MetadataControl(BaseControl):
         datafile = populate.add_mutually_exclusive_group()
         datafile.add_argument("--file", help="Input file")
         datafile.add_argument(
-            "--fileid", type=long, help="Input OriginalFile ID")
+            "--fileid", type=int, help="Input OriginalFile ID")
 
         cfgfile = populate.add_mutually_exclusive_group()
         cfgfile.add_argument("--cfg", help="YAML configuration file")
         cfgfile.add_argument(
-            "--cfgid", type=long, help="YAML configuration OriginalFile ID")
+            "--cfgid", type=int, help="YAML configuration OriginalFile ID")
 
         populate.add_argument("--attach", action="store_true", help=(
             "Upload input or configuration files and attach to parent object"))
